add story for EventDateRange for explanation
