+++ common/views/components/EventDateRange/README.md
@@ -1,3 +1,12 @@
-We select the earliest, future date to display on an event.
+# What's the Purpose?
+To be able to express a list of date ranges as a single, relevant date range
+for people to be able to get a snapshot of an event.
+
+# How do we do this?
+We display the earliest, future date, which would be most relevant to a user
+trying to make a decision on whether they would like to attend an event.
 
 If this isn't available, we use the latest possible date.
+
+This is usually used in conjunction with a link to view the full list of date
+ranges.
